requirements.txt
openai
panel
python-dotenv